sound waveform improved

+++ src/main/java/org/kdzumba/graphics2d/Coordinate.java
@@ -1,15 +1,15 @@
 package org.kdzumba.graphics2d;
 
 public class Coordinate {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
     public Coordinate() {
-        this.x = 0;
-        this.y = 0;
+        this.x = 0.0;
+        this.y = 0.0;
     }
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
